pyproject.toml
[project]
name = "monmon"
version = "0.1.0"
description = "Monitor any website and record its metrics to database"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT License"}
authors = [
  {name = "Valerii Krivchikov", email = "tehbrut@gmail.com"},
]
maintainers = [
  {name = "Valerii Krivchikov", email = "tehbrut@gmail.com"}
]
keywords = ["tracking", "monitoring", "web"]
dependencies = [
    "psycopg2>=2.9.6",
    "types-psycopg2>=2.9",
    "aiohttp[speedups]>=3.8.4",
    "yoyo-migrations>=8.2.0",
    "structlog>=23.1.0",
    "validators>=0.20.0",
    "pytest>=7.3.1",
]
[project.optional-dependencies]
linters=[
    "pylint[spelling]>=2.17.3",
    "mypy>=1.2.0",
    "black>=23.3.0",
]
dynamic = ["version", "description", "optional-dependencies"]

[build-system]
requires = ["setuptools>=65.5.1"]
build-backend = "setuptools.build_meta"

[[tool.mypy.overrides]]
module = ["yoyo.*", "validators"]
ignore_missing_imports = true

[tool.pylint.'MASTER']
ignore="""
    migrations,
    config.ini,
    venv,
    requirements.txt,
    requirements-dev.txt,
    pyproject.toml,
    LICENSE,
    README.md,
    """
[tool.pylint.'FORMAT']
max-line-length=150

[tool.pylint.'MESSAGE_CONTROL']
disable="""
    too-few-public-methods,
    """